lib/chaincode/utils_test.go
package chaincode // Replace with your actual package name

import (
	"encoding/json"
	"pedersen-commitment-transfer/lib/tests/testsfakes"
	"testing"

	"github.com/bwesterb/go-ristretto"
	"github.com/golang/protobuf/ptypes/timestamp"
	"github.com/hyperledger/fabric-chaincode-go/shim"
	"github.com/stretchr/testify/assert"
)

func TestGetBlockNumber(t *testing.T) {
	// Create a mock stub
	ctx := &testsfakes.FakeTestTransactionContextInterface{}
	stub := &testsfakes.FakeTestChaincodeStubInterface{}
	ctx.GetStubStub = func() shim.ChaincodeStubInterface {
		return stub
	}
	expectedTimestamp := &timestamp.Timestamp{
		Seconds: 123456789, // Replace with your desired timestamp
	}

	// Configure the fake stub's behavior
	stub.GetTxTimestampReturns(expectedTimestamp, nil)

	// Call your function with the fake stub
	blockNumber, err := GetBlockNumber(stub)

	// Assertions
	assert.NoError(t, err)
	assert.Equal(t, int64(12345678), blockNumber) // Replace with your expected block number
}

func TestCreateTxInfo(t *testing.T) {
	// Create a mock stub
	ctx := &testsfakes.FakeTestTransactionContextInterface{}
	stub := &testsfakes.FakeTestChaincodeStubInterface{}
	ctx.GetStubStub = func() shim.ChaincodeStubInterface {
		return stub
	}

	// Define the expected timestamp
	expectedTimestamp := &timestamp.Timestamp{
		Seconds: 123456789, // Replace with your desired timestamp
	}

	// Configure the fake stub's behavior for GetTxTimestamp
	stub.GetTxTimestampReturns(expectedTimestamp, nil)

	// Call your function with the fake stub
	sender := "sender"
	amount := ristretto.Point{} // Replace with your desired amount
	txInfo, err := createTxInfo(stub, sender, amount)

	// Custom assertions
	if err != nil {
		t.Errorf("createTxInfo error: %v", err)
	}
	if txInfo.Sender != sender {
		t.Errorf("createTxInfo Sender: expected %s, got %s", sender, txInfo.Sender)
	}
	// Add more custom assertions as needed for other fields
}

func TestStoreTxInfo(t *testing.T) {
	// Create a mock stub
	ctx := &testsfakes.FakeTestTransactionContextInterface{}
	stub := &testsfakes.FakeTestChaincodeStubInterface{}
	ctx.GetStubStub = func() shim.ChaincodeStubInterface {
		return stub
	}

	// Define the expected timestamp
	expectedTimestamp := &timestamp.Timestamp{
		Seconds: 123456789, // Replace with your desired timestamp
	}

	// Configure the fake stub's behavior for GetTxTimestamp
	stub.GetTxTimestampReturns(expectedTimestamp, nil)

	// Call your function with the fake stub
	sender := "sender"
	amount := ristretto.Point{} // Replace with your desired amount
	err := storeTxInfo(stub, sender, amount)

	// Custom assertions
	if err != nil {
		t.Errorf("storeTxInfo error: %v", err)
	}
	// Add more custom assertions as needed
}

func TestGetTxInfo(t *testing.T) {
	// Create a mock stub
	ctx := &testsfakes.FakeTestTransactionContextInterface{}
	stub := &testsfakes.FakeTestChaincodeStubInterface{}
	ctx.GetStubStub = func() shim.ChaincodeStubInterface {
		return stub
	}

	// Define the expected timestamp
	expectedTimestamp := &timestamp.Timestamp{
		Seconds: 123456789, // Replace with your desired timestamp
	}

	// Configure the fake stub's behavior for GetTxTimestamp
	stub.GetTxTimestampReturns(expectedTimestamp, nil)

	// Create a sample TxInformation struct and store it in the stub's state
	sender := "sender"
	amount := ristretto.Point{} // Replace with your desired amount
	txInfo := TxInformation{
		Sender:              sender,
		Amount:              amount.Bytes(),
		ProposalBlockNumber: 12345678, // Replace with your desired block number
		isValid:             true,
	}
	txInfoBytes, _ := json.Marshal(txInfo)
	stub.GetStateReturns(txInfoBytes, nil)

	// Call your function with the fake stub
	txId := "tx123" // Replace with your desired transaction ID
	resultTxInfo, err := getTxInfo(stub, txId)

	// Custom assertions
	if err != nil {
		t.Errorf("getTxInfo error: %v", err)
	}
	if resultTxInfo.Sender != sender {
		t.Errorf("getTxInfo Sender: expected %s, got %s", sender, resultTxInfo.Sender)
	}
	// Add more custom assertions as needed for other fields
}
